added signout api and admin dashboard

+++ src/actions/auth.actions.js
@@ -1,3 +1,4 @@
+import axios from "axios"
 import axiosInstance from "../helpers/axios"
 import { authConstants } from "./constants"
 
@@ -66,9 +67,24 @@ export const isUserLoggedIn = () => {
 
 export const Signout = () => {
     return async dispatch => {
-        localStorage.clear();
+
         dispatch({
             type: authConstants.LOGOUT_REQUEST
         })
+        const res = await axiosInstance.post(`/admin/signout`);
+
+        if (res.status === 200) {
+            localStorage.clear()
+            dispatch({
+                type: authConstants.LOGOUT_SUCCESS
+            });
+        } else {
+            dispatch({
+                type: authConstants.LOGOUT_FAILURE,
+                payload: {
+                    error: res.data.error
+                }
+            })
+        }
     }
 }
